+++ demo/index.js
@@ -50,10 +50,10 @@ if (window.innerWidth < 768) {
 }
 
 window.vditor = new Vditor('vditor', {
-  _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
-  // _lutePath: 'src/js/lute/lute.min.js',
+  // _lutePath: `http://192.168.0.107:9090/lute.min.js?${new Date().getTime()}`,
+  _lutePath: 'src/js/lute/lute.min.js',
   toolbar,
-  mode: 'ir',
+  mode: 'wysiwyg',
   height: window.innerHeight + 100,
   outline: true,
   debugger: true,
@@ -63,6 +63,7 @@ window.vditor = new Vditor('vditor', {
     markdown: {
       toc: true,
       mark: true,
+      footnotes: true,
     },
   },
   toolbarConfig: {