+++ src/index.ts
@@ -32,6 +32,7 @@ import {Options} from "./ts/util/Options";
 import {processCodeRender} from "./ts/util/processCode";
 import {getCursorPosition, getEditorRange} from "./ts/util/selection";
 import {WYSIWYG} from "./ts/wysiwyg";
+import {afterRenderEvent} from "./ts/wysiwyg/afterRenderEvent";
 import {input} from "./ts/wysiwyg/input";
 import {renderDomByMd} from "./ts/wysiwyg/renderDomByMd";
 
@@ -356,6 +357,97 @@ class Vditor extends VditorMethod {
         this.vditor.element.removeAttribute("style");
         document.getElementById("vditorIconScript").remove();
         this.clearCache();
+    }
+
+    /** 获取评论 ID */
+    public getCommentIds() {
+        if (this.vditor.currentMode !== "wysiwyg") {
+            return [];
+        }
+        return this.vditor.wysiwyg.getComments(this.vditor, true);
+    }
+
+    /** 高亮评论 */
+    public hlCommentIds(ids: string[]) {
+        if (this.vditor.currentMode !== "wysiwyg") {
+            return;
+        }
+        const hlItem = (item: Element) => {
+            item.classList.remove("vditor-comment--hover");
+            ids.forEach((id) => {
+                if (item.getAttribute("data-cmtids").indexOf(id) > -1) {
+                    item.classList.add("vditor-comment--hover");
+                }
+            });
+        };
+        this.vditor.wysiwyg.element.querySelectorAll(".vditor-comment").forEach((item) => {
+            hlItem(item);
+        });
+        if (this.vditor.preview.element.style.display !== "none") {
+            this.vditor.preview.element.querySelectorAll(".vditor-comment").forEach((item) => {
+                hlItem(item);
+            });
+        }
+    }
+
+    /** 取消评论高亮 */
+    public unHlCommentIds(ids: string[]) {
+        if (this.vditor.currentMode !== "wysiwyg") {
+            return;
+        }
+        const unHlItem = (item: Element) => {
+            ids.forEach((id) => {
+                if (item.getAttribute("data-cmtids").indexOf(id) > -1) {
+                    item.classList.remove("vditor-comment--hover");
+                }
+            });
+        };
+        this.vditor.wysiwyg.element.querySelectorAll(".vditor-comment").forEach((item) => {
+            unHlItem(item);
+        });
+        if (this.vditor.preview.element.style.display !== "none") {
+            this.vditor.preview.element.querySelectorAll(".vditor-comment").forEach((item) => {
+                unHlItem(item);
+            });
+        }
+    }
+
+    /** 删除评论 */
+    public removeCommentIds(removeIds: string[]) {
+        if (this.vditor.currentMode !== "wysiwyg") {
+            return;
+        }
+
+        const removeItem = (item: Element, removeId: string) => {
+            const ids = item.getAttribute("data-cmtids").split(" ");
+            ids.find((id, index) => {
+                if (id === removeId) {
+                    ids.splice(index, 1);
+                    return true;
+                }
+            });
+            if (ids.length === 0) {
+                item.outerHTML = item.innerHTML;
+                getEditorRange(this.vditor.element).collapse(true);
+            } else {
+                item.setAttribute("data-cmtids", ids.join(" "));
+            }
+        };
+        removeIds.forEach((removeId) => {
+            this.vditor.wysiwyg.element.querySelectorAll(".vditor-comment").forEach((item) => {
+                removeItem(item, removeId);
+            });
+            if (this.vditor.preview.element.style.display !== "none") {
+                this.vditor.preview.element.querySelectorAll(".vditor-comment").forEach((item) => {
+                    removeItem(item, removeId);
+                });
+            }
+        });
+        afterRenderEvent(this.vditor, {
+            enableAddUndoStack: true,
+            enableHint: false,
+            enableInput: false,
+        });
     }
 }