+++ src/ts/util/editorCommonEvent.ts
@@ -6,14 +6,14 @@ import {processKeydown as mdProcessKeydown} from "../sv/processKeydown";
 import {setEditMode} from "../toolbar/EditMode";
 import {hidePanel} from "../toolbar/setToolbar";
 import {uploadFiles} from "../upload";
-import {getCursorPosition} from "../util/selection";
 import {afterRenderEvent} from "../wysiwyg/afterRenderEvent";
 import {processKeydown} from "../wysiwyg/processKeydown";
 import {removeHeading, setHeading} from "../wysiwyg/setHeading";
 import {getEventName, isCtrl} from "./compatibility";
 import {getSelectText} from "./getSelectText";
-import {hasClosestByMatchTag} from "./hasClosest";
+import {hasClosestByAttribute, hasClosestByMatchTag} from "./hasClosest";
 import {matchHotKey} from "./hotKey";
+import {getCursorPosition} from "./selection";
 
 export const focusEvent = (vditor: IVditor, editorElement: HTMLElement) => {
     editorElement.addEventListener("focus", () => {
@@ -25,12 +25,15 @@ export const focusEvent = (vditor: IVditor, editorElement: HTMLElement) => {
 };
 
 export const blurEvent = (vditor: IVditor, editorElement: HTMLElement) => {
-    editorElement.addEventListener("blur", () => {
+    editorElement.addEventListener("blur", (event) => {
         if (vditor.currentMode === "ir") {
             const expandElement = vditor.ir.element.querySelector(".vditor-ir__node--expand");
             if (expandElement) {
                 expandElement.classList.remove("vditor-ir__node--expand");
             }
+        } else if (vditor.currentMode === "wysiwyg" &&
+            !vditor.wysiwyg.selectPopover.contains(event.relatedTarget as HTMLElement)) {
+            vditor.wysiwyg.hideComment();
         }
         if (vditor.options.blur) {
             vditor.options.blur(getMarkdown(vditor));
@@ -63,11 +66,18 @@ export const cutEvent =
     (vditor: IVditor, editorElement: HTMLElement, copy: (event: ClipboardEvent, vditor: IVditor) => void) => {
         editorElement.addEventListener("cut", (event: ClipboardEvent) => {
             copy(event, vditor);
+            // 获取 comment
+            if (vditor.options.comment.enable && vditor.currentMode === "wysiwyg") {
+                vditor.wysiwyg.getComments(vditor);
+            }
             document.execCommand("delete");
         });
     };
 
 export const scrollCenter = (vditor: IVditor) => {
+    if (vditor.currentMode === "wysiwyg" && vditor.options.comment.enable) {
+        vditor.options.comment.adjustTop(vditor.wysiwyg.getComments(vditor, true));
+    }
     if (!vditor.options.typewriterMode) {
         return;
     }
@@ -88,6 +98,12 @@ export const hotkeyEvent = (vditor: IVditor, editorElement: HTMLElement) => {
         if ((vditor.options.hint.extend.length > 1 || vditor.toolbar.elements.emoji) &&
             vditor.hint.select(event, vditor)) {
             return;
+        }
+
+        // 获取 comment
+        if (vditor.options.comment.enable && vditor.currentMode === "wysiwyg" &&
+            (event.key === "Backspace" || matchHotKey("⌘-X", event))) {
+            vditor.wysiwyg.getComments(vditor);
         }
 
         if (vditor.currentMode === "sv") {
@@ -195,15 +211,28 @@ export const hotkeyEvent = (vditor: IVditor, editorElement: HTMLElement) => {
 };
 
 export const selectEvent = (vditor: IVditor, editorElement: HTMLElement) => {
-    if (!vditor.options.select) {
-        return;
-    }
-    editorElement.addEventListener("selectstart", (event: Event) => {
+    editorElement.addEventListener("selectstart", (event: Event & { target: HTMLElement }) => {
         editorElement.onmouseup = () => {
-            const selectText = getSelectText(vditor[vditor.currentMode].element);
-            if (selectText) {
-                vditor.options.select(selectText);
-            }
+            setTimeout(() => { // 鼠标放开后 range 没有即时更新
+                const selectText = getSelectText(vditor[vditor.currentMode].element);
+                if (selectText.trim()) {
+                    if (vditor.currentMode === "wysiwyg" && vditor.options.comment.enable) {
+                        if (!hasClosestByAttribute(event.target, "data-type", "footnotes-block") &&
+                            !hasClosestByAttribute(event.target, "data-type", "link-ref-defs-block")) {
+                            vditor.wysiwyg.showComment();
+                        } else {
+                            vditor.wysiwyg.hideComment();
+                        }
+                    }
+                    if (vditor.options.select) {
+                        vditor.options.select(selectText);
+                    }
+                } else {
+                    if (vditor.currentMode === "wysiwyg" && vditor.options.comment.enable) {
+                        vditor.wysiwyg.hideComment();
+                    }
+                }
+            });
         };
     });
 };